fix(abi): NodoJson int/float → int64_t/double (cumple Manual 2 §4.1)
HALLAZGO: synapse_rt_types.h usaba int/float (56 bytes) para NodoJson
pero Manual 2 §4.1 dice que entero=int64_t y decimal=double.
Esto creaba ABI mismatch con el generador S1 que emitía int64_t/double.

FIX:
- synapse_rt_types.h: NodoJson tipo/valor_bool/longitud → int64_t, valor_num → double
- synapse_rt_types.h: NodoArr/ParArr count/cap → int64_t
- runtime/core/io.c: stdin en modo binario para LSP (evita \r\n en Windows)
- tests/unit/test_nodojson_abi.c: TDD verifica sizeof(NodoJson)==64 y offsets

VERIFICACIÓN:
- sizeof(NodoJson) == 64 bytes ✓
- Tests TDD: 4/4 PASS
- Python tests: 17/17 PASS
- Auditoría: 0 brechas

DEUDA PENDIENTE: LSP solo procesa 1 mensaje (pre-existente, no causado por este fix)

+++ synapse_rt_types.h
@@ -284,26 +284,26 @@ struct ParJson {
 };
 
 struct NodoJson {
-    int tipo;                // -1=Error, 0=Nulo, 1=Booleano, 2=Numero, 3=Cadena, 4=Arreglo, 5=Objeto
-    int valor_bool;
-    float valor_num;
+    int64_t tipo;           // cumple Manual 2 §4.1: entero = int64_t
+    int64_t valor_bool;     // cumple Manual 2 §4.1: entero = int64_t
+    double valor_num;       // cumple Manual 2 §4.1: decimal = double
     CadenaSegura valor_str;
     NodoJson* arreglo_hijos;
     ParJson* objeto_pares;
-    int longitud;
+    int64_t longitud;       // cumple Manual 2 §4.1: entero = int64_t
 };
 
 // --- Dynamic array helpers for JSON ---
 typedef struct {
     NodoJson* items;
-    int count;
-    int cap;
+    int64_t count;          // cumple Manual 2 §4.1: entero = int64_t
+    int64_t cap;            // cumple Manual 2 §4.1: entero = int64_t
 } NodoArr;
 
 typedef struct {
     ParJson* items;
-    int count;
-    int cap;
+    int64_t count;          // cumple Manual 2 §4.1: entero = int64_t
+    int64_t cap;            // cumple Manual 2 §4.1: entero = int64_t
 } ParArr;
 
 // --- Watchdog types (debug mode) ---
